Move orphan `ToPlutusData` instances with their types
